+++ src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomepageComponent } from './homepage/homepage.component';
-import { ProjectsComponent } from './projects/projects.component';
-import { ProblemsComponent } from './problems/problems.component';
-import { BlogComponent } from './blog/blog.component';
-import { AboutComponent } from './about/about.component';
-import { ProblemMdDisplayComponent } from './problem-md-display/problem-md-display.component';
-import { ArticleMdDisplayComponent } from './article-md-display/article-md-display.component';
+import { ProjectsComponent } from './base/projects/projects.component';
+import { ProblemsComponent } from './base/problems/problems.component';
+import { BlogComponent } from './base/blog/blog.component';
+import { AboutComponent } from './base/about/about.component';
+import { ProblemMdDisplayComponent } from './markdown/problem-md-display/problem-md-display.component';
+import { ArticleMdDisplayComponent } from './markdown/article-md-display/article-md-display.component';
 import { TestComponent } from './test/test.component';
 
 const routes: Routes = [
