+++ src/app/app.module.ts
@@ -7,15 +7,15 @@ import { FormsModule } from '@angular/forms';
 import { TopNavbarComponent } from './top-navbar/top-navbar.component';
 import { AppRoutingModule } from './app-routing.module';
 import { HomepageComponent } from './homepage/homepage.component';
-import { ProjectsComponent } from './projects/projects.component';
-import { BlogComponent } from './blog/blog.component';
-import { AboutComponent } from './about/about.component';
-import { ProblemsComponent } from './problems/problems.component';
+import { ProjectsComponent } from './base/projects/projects.component';
+import { BlogComponent } from './base/blog/blog.component';
+import { AboutComponent } from './base/about/about.component';
+import { ProblemsComponent } from './base/problems/problems.component';
 import { MarkdownModule } from 'ngx-markdown';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
-import { MdDisplayComponent } from './md-display/md-display.component';  
-import { ProblemMdDisplayComponent } from './problem-md-display/problem-md-display.component';
-import { ArticleMdDisplayComponent } from './article-md-display/article-md-display.component';  
+import { MdDisplayComponent } from './markdown/md-display/md-display.component';  
+import { ProblemMdDisplayComponent } from './markdown/problem-md-display/problem-md-display.component';
+import { ArticleMdDisplayComponent } from './markdown/article-md-display/article-md-display.component';  
 import { NgxMathMdModule } from 'projects/ngx-math-md/src/public-api';
 import { TestComponent } from './test/test.component';
