+++ src/app/base/blog/blog.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { MdService } from '../md.service';
-import { Article } from '../model/Article';
+import { MdService } from '../../md.service';
+import { Article } from '../../model/Article';
 
 
 @Component({
