+++ src/app/base/problems/problems.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { MdService } from '../md.service';
-import { Problem } from '../model/Problem';
+import { MdService } from '../../md.service';
+import { Problem } from '../../model/Problem';
 
 @Component({
   selector: 'app-problems',
