+++ src/app/markdown/problem-md-display/problem-md-display.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { MdService } from '../md.service';
-import { Problem } from '../model/Problem';
+import { MdService } from '../../md.service';
+import { Problem } from '../../model/Problem';
 
 
 @Component({
