+++ src/app/model/problems.ts
@@ -1,5 +1,5 @@
 import { Problem } from './Problem';
 
 export const PROBLEMS: Problem[] = [
-    { id: 1, title:"Combine to 61A", path: "assets/problems_md/combine61a.md"}
+    { id: 1, title:"Combine to 61A", path: "assets/md/problems/combine61a.md"}
 ]
