+++ src/app/test/test.component.html
@@ -1,3 +1,3 @@
 <div class="container">
-    <ngx-math-md [src]=mdPath></ngx-math-md>
+    <markdown [src]=mdPath></markdown>
 </div>
