+++ src/assets/pdf/PipelinedJoins.pdf
